Squashed commit of the following:
    fix notes parsing for metron and many variations
    move notes parsing into another file.
    add comicinfo metron origin test
    rename modules to not shadow python builtins
    fix binary pdf files for new mupdf
