+++ comicbox/schemas/comictagger.py
@@ -6,14 +6,14 @@ from types import MappingProxyType
 
 from marshmallow.fields import Nested
 
-from comicbox.fields.collections import StringListField, StringSetField
-from comicbox.fields.enum import AgeRatingField, MangaField, YesNoField
+from comicbox.fields.collection_fields import StringListField, StringSetField
+from comicbox.fields.enum_fields import AgeRatingField, MangaField, YesNoField
 from comicbox.fields.fields import StringField
-from comicbox.fields.numbers import BooleanField, DecimalField, IntegerField
+from comicbox.fields.number_fields import BooleanField, DecimalField, IntegerField
 from comicbox.fields.pycountry import CountryField, LanguageField
 from comicbox.schemas.base import BaseSchema, BaseSubSchema
 from comicbox.schemas.comicbookinfo import ComicBookInfoCreditSchema
-from comicbox.schemas.json import JsonSchema, JsonSubSchema
+from comicbox.schemas.json_schemas import JsonSchema, JsonSubSchema
 
 TAG_ORIGIN_KEY = "tag_origin"
 ISSUE_ID_KEY = "issue_id"
@@ -101,7 +101,7 @@ class ComictaggerSubSchema(JsonSubSchema):
     pages = Nested(ComictaggerPageInfoSchema, many=True)
 
     # comet unique
-    price = DecimalField(minimum=Decimal(0.0))
+    price = DecimalField(minimum=Decimal("0.0"))
     is_version_of = StringSetField(as_string=True)
     rights = StringField()
     identifier = StringSetField(as_string=True)
