+++ comicbox/transforms/comicbox_json.py
@@ -2,7 +2,7 @@
 
 from comicbox.schemas.comicbox_json import ComicboxJsonSchema
 from comicbox.transforms.comicbox_mixin import ComicboxTransformMixin
-from comicbox.transforms.json import JsonTransform
+from comicbox.transforms.json_transforms import JsonTransform
 
 
 class ComicboxJsonTransform(JsonTransform, ComicboxTransformMixin):
