+++ comicbox/transforms/comicinfo_storyarcs.py
@@ -3,7 +3,7 @@
 from itertools import zip_longest
 from logging import getLogger
 
-from comicbox.fields.numbers import IntegerField
+from comicbox.fields.number_fields import IntegerField
 from comicbox.schemas.comicbox_mixin import STORY_ARCS_KEY
 from comicbox.transforms.base import BaseTransform
 
@@ -38,7 +38,7 @@ class ComicInfoStoryArcsTransformMixin(BaseTransform):
                     number = integer_field.deserialize(number_str)
             except Exception:
                 LOG.exception(
-                    f"{self._path}:" f" Deserialize story_arc_number{name}:{number_str}"
+                    f"{self._path}: Deserialize story_arc_number{name}:{number_str}"
                 )
                 number = None
             story_arcs[name] = number
