+++ comicbox/transforms/pdf.py
@@ -8,7 +8,7 @@ from xml.sax.saxutils import unescape
 from bidict import bidict
 
 from comicbox.dict_funcs import deep_update
-from comicbox.fields.collections import StringSetField
+from comicbox.fields.collection_fields import StringSetField
 from comicbox.schemas.comicbox_mixin import (
     CONTRIBUTORS_KEY,
     GENRES_KEY,
@@ -33,7 +33,7 @@ from comicbox.transforms.comicinfo import ComicInfoTransform
 from comicbox.transforms.comictagger import ComictaggerTransform
 from comicbox.transforms.filename import FilenameTransform
 from comicbox.transforms.metroninfo import MetronInfoTransform
-from comicbox.transforms.xml import XmlTransform
+from comicbox.transforms.xml_transforms import XmlTransform
 
 _KEYWORDS_TRANSFORM_CLASSES = (
     # Different order than all sources
