+++ package.json
@@ -35,17 +35,17 @@
   },
   "devDependencies": {
     "@eslint-community/eslint-plugin-eslint-comments": "^4.4.1",
-    "@eslint/json": "^0.8.0",
-    "@fsouza/prettierd": "^0.25.1",
+    "@eslint/json": "^0.9.0",
+    "@fsouza/prettierd": "^0.26.0",
     "@prettier/plugin-xml": "^3.1.0",
+    "@stylistic/eslint-plugin": "^2.13.0",
     "eslint": "^9.16.0",
-    "eslint-config-prettier": "^9.0.0",
+    "eslint-config-prettier": "^10.0.1",
     "eslint-import-resolver-alias": "^1.1.2",
     "eslint-plugin-array-func": "^5.0.1",
     "eslint-plugin-compat": "^6.0.0",
     "eslint-plugin-depend": "^0.12.0",
     "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-json-schema-validator": "^5.1.3",
     "eslint-plugin-mdx": "^3.1.5",
     "eslint-plugin-no-secrets": "^2.1.1",
     "eslint-plugin-no-unsanitized": "^4.1.2",
