+++ pyproject.toml
@@ -3,7 +3,7 @@ name = "comicbox"
 requires-python = ">=3.10"
 
 [build-system]
-requires = ["poetry-core>=1.0.0"]
+requires = ["poetry-core>=2.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
@@ -60,11 +60,11 @@ xmltodict = "^0.14.0"
 wheel = "^0.45.1"
 
 [tool.poetry.group.lint.dependencies]
-check-jsonschema = "^0.30.0"
+check-jsonschema = "^0.31.0"
 codespell = "^2.1.0"
 pyright = "^1.1.354"
 radon = { version = "^6.0.1", extras = ["toml"] }
-ruff = "^0.8.1"
+ruff = "^0.9.1"
 types-python-dateutil = "^2.8.19.13"
 vulture = "^2.1"
