+++ tests/schemas/test_json.py
@@ -7,7 +7,7 @@ from types import MappingProxyType
 
 import simplejson as json
 
-from comicbox.fields.enum import PageTypeEnum
+from comicbox.fields.enum_fields import PageTypeEnum
 from comicbox.schemas.comicbox_json import ComicboxJsonSchema
 from comicbox.schemas.comicbox_mixin import ROOT_TAG
 from comicbox.transforms.comicbox_json import ComicboxJsonTransform
