+++ tests/schemas/test_yaml.py
@@ -5,7 +5,7 @@ from io import StringIO
 from pathlib import Path
 from types import MappingProxyType
 
-from comicbox.fields.enum import PageTypeEnum
+from comicbox.fields.enum_fields import PageTypeEnum
 from comicbox.schemas.comicbox_mixin import ROOT_TAG
 from comicbox.schemas.comicbox_yaml import ComicboxYamlSchema
 from comicbox.schemas.yaml import YamlRenderModule
